Código foi ajustado

+++ src/main/java/com/example/mercadinho/controller/request/ProductRequest.java
@@ -2,11 +2,13 @@ package com.example.mercadinho.controller.request;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record ProductRequest(
         String id,
         String name,
-        Float price,
-        Long quantity
+        BigDecimal price,
+        Integer quantity
 ) {
 }
