+++ src/main/java/com/example/mercadinho/controller/response/ProductResonse.java
@@ -3,10 +3,12 @@ package com.example.mercadinho.controller.response;
 import com.example.mercadinho.domain.repository.model.ProductEntity;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record ProductResonse(
         String name,
-        Float price
+        BigDecimal price
 ) {
     public ProductEntity fromEntity(ProductEntity productEntity) {
         return ProductEntity.builder()
